add the '--no-fcs-nack' option
When a corrupted frame is received (bad FCS check), the default
behaviour is to send back an NACK with the sequence number contained in
the received frame. The '--no-fcs-nack' option allows to change this
behaviour. If this last is enabled, no NACK will be sent back.
Consequently, we let the sender of the corrupted frame send it back
after its timeout.